requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2==2.9.6
python-decouple==3.8
python-dotenv==1.0.0
drf-spectacular==0.26.2